+++ HONBAB/src/main/webapp/WEB-INF/views/admin/ADMIndex.jsp
@@ -16,7 +16,7 @@
 
 <div class="cd1">
 ${loginUser} 님 관리자로 로그인하셨습니다.
-<a id="logout" href="${contextPath }/oper/logout">Logout</a>
+<a id="logout" href="${contextPath }/member/logout">Logout</a>
   <h1>어드민홈</h1>
   <h2>제목</h2>
   <p>내용</p>
